+++ CHANGELOG.md
@@ -16,7 +16,7 @@ Since version 1.0.0, we try to follow the [Semantic Versioning](https://semver.o
 
 - Processing JSON availability payload should not result in a TypeError anymore.
 - Minor changes to be compatible with the upcoming Homebridge v2 release, amongst others:
-  - In Light accessory, set default value before configuring min/max values.
+  - In most services where the (numeric) range of a characteristic is limited, the value is now set correctly before doing so, to prevent warnings from HAP-NodeJS.
   - Sanitize accessory names so they only contain alphanumeric, space, and apostrophe characters, and start with an alphanumeric character.
 
 ## [1.11.0-beta.7] - 2025-01-04