+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 Since version 1.0.0, we try to follow the [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard.
 
-## [Unreleased]
+## [1.11.0-beta.8] - 2025-01-04
 
 ### Changed
 
@@ -474,7 +474,8 @@ For `cover` devices the following changes/fixes are in this release:
 - Restore BatteryService and WindowCovering properly on start up.
 - Improve state determination for WindowCovering.
 
-[unreleased]: https://github.com/itavero/homebridge-z2m/compare/v1.11.0-beta.7...HEAD
+[unreleased]: https://github.com/itavero/homebridge-z2m/compare/v1.11.0-beta.8...HEAD
+[1.11.0-beta.8]: https://github.com/itavero/homebridge-z2m/compare/v1.11.0-beta.7...v1.11.0-beta.8
 [1.11.0-beta.7]: https://github.com/itavero/homebridge-z2m/compare/v1.11.0-beta.6...v1.11.0-beta.7
 [1.11.0-beta.6]: https://github.com/itavero/homebridge-z2m/compare/v1.11.0-beta.5...v1.11.0-beta.6
 [1.11.0-beta.5]: https://github.com/itavero/homebridge-z2m/compare/v1.11.0-beta.4...v1.11.0-beta.5